MFH [rev 1.8]: SIlence a type size mismatch warning.
git-svn-id: https://vrjuggler.googlecode.com/svn/branches/VRJ_RELENG_2_0_BRANCH/juggler@17698 08b38cba-cd3b-11de-854e-f91c5b6e4272

+++ modules/gadgeteer/gadget/AbstractNetworkManager.h
@@ -136,7 +136,7 @@ public:
    /**
     * Get the number of nodes in network.
     */
-   int getNumNodes()
+   size_t getNumNodes()
    {
       return mNodes.size();
    }
